requirements.txt
flask>=2.3.0
opencv-contrib-python>=4.8.0
# Pin ML stack to compatible Windows wheels
scikit-learn==1.3.2
numpy==1.26.4
scipy==1.10.1
Werkzeug>=2.3.0
Pillow>=10.0.0
flask-login>=0.6.0
pymongo>=4.6.0
dnspython>=2.4.0
python-dotenv>=1.0.0
